Faster caching

+++ automl/components/transformers/utils.py
@@ -2,8 +2,8 @@ import numpy as np
 
 from .transformer import DataType
 
-
-def categorical_column_to_int_categories(col, int_dtype: type = np.int32):
+# TODO: look into codes
+def categorical_column_to_int_categories(col, int_dtype: type = np.uint8):
     if DataType.is_categorical(col.dtype):
         col.cat.categories = np.arange(
             start=0, stop=len(col.cat.categories), dtype=int_dtype
